+++ macosx/Badger.m
@@ -72,8 +72,8 @@
 }
 
 - (void) updateBadgeWithCompleted: (int) completed
-                    uploadRate: (int) uploadRate
-                    downloadRate: (int) downloadRate
+                    uploadRate: (float) uploadRate
+                    downloadRate: (float) downloadRate
 {
     NSImage * dockIcon = nil;
     NSSize iconSize = [fDockIcon size];
